Factored out the core SoftAP code to make it Browserify-friendly. SoftAP can now be run in the browser. Updated README.md with instructions. Tests pass. No change to repl or existing usage.
